+++ client/src/BibleEngineClient.ts
@@ -1,10 +1,11 @@
-import { ConnectionOptions } from 'typeorm';
+import { DataSourceOptions } from 'typeorm';
 
 import {
     BibleEngine,
     BibleEngineOptions,
     BibleVersionRemoteOnlyError,
     IBibleReferenceRangeQuery,
+    IBibleSearchOptions,
 } from '@bible-engine/core';
 import { BibleApi } from './Bible.api';
 
@@ -17,7 +18,7 @@ export class BibleEngineClient {
         bibleEngineOptions,
         apiBaseUrl,
     }: {
-        bibleEngineConnectionOptions?: ConnectionOptions;
+        bibleEngineConnectionOptions?: DataSourceOptions;
         bibleEngineOptions?: BibleEngineOptions;
         apiBaseUrl?: string;
     }) {
@@ -79,8 +80,37 @@ export class BibleEngineClient {
             }
         }
 
-        if (this.remoteApi)
-            return this.remoteApi.getReferenceRange(rangeQuery).then((resp) => resp.result);
+        if (this.remoteApi) {
+            // we use cachable endpoints for the request:
+            // single chapter
+            if (
+                rangeQuery.versionChapterNum &&
+                (!rangeQuery.versionChapterEndNum ||
+                    rangeQuery.versionChapterEndNum === rangeQuery.versionChapterNum) &&
+                !rangeQuery.versionVerseNum &&
+                !rangeQuery.versionVerseEndNum
+            )
+                return this.remoteApi
+                    .getChapter({
+                        versionUid: rangeQuery.versionUid,
+                        osisId: rangeQuery.bookOsisId,
+                        chapterNr: rangeQuery.versionChapterNum,
+                    })
+                    .then((resp) => resp.result);
+            // one or multiple verses
+            else if (rangeQuery.versionChapterNum && rangeQuery.versionVerseNum)
+                return this.remoteApi
+                    .getVerses({
+                        versionUid: rangeQuery.versionUid,
+                        osisId: rangeQuery.bookOsisId,
+                        chapterNr: rangeQuery.versionChapterNum,
+                        verseNr: rangeQuery.versionVerseNum,
+                        chapterEnd: rangeQuery.versionChapterEndNum,
+                        verseEnd: rangeQuery.versionVerseEndNum,
+                    })
+                    .then((resp) => resp.result);
+            else return this.remoteApi.getReferenceRange(rangeQuery).then((resp) => resp.result);
+        }
         throw new Error(`can't get formatted text: invalid version`);
     }
 
@@ -100,5 +130,26 @@ export class BibleEngineClient {
             if (this.remoteApi) return this.remoteApi.getVersions().then((resp) => resp.result);
             throw new Error(`No remote config provided`);
         } else return this.localBibleEngine.getVersions();
+    }
+
+    search(searchOptions: IBibleSearchOptions, forceRemote = false) {
+        if (!this.localBibleEngine || forceRemote) {
+            if (this.remoteApi)
+                return this.remoteApi
+                    .search({
+                        ...searchOptions,
+                        page: searchOptions.pagination ? searchOptions.pagination.page : undefined,
+                        count: searchOptions.pagination
+                            ? searchOptions.pagination.count
+                            : undefined,
+                        rangeStart: searchOptions.bookRange
+                            ? searchOptions.bookRange.start
+                            : undefined,
+                        rangeEnd: searchOptions.bookRange ? searchOptions.bookRange.end : undefined,
+                    })
+                    .then((resp) => resp.result);
+            throw new Error(`No remote config provided`);
+        }
+        return this.localBibleEngine.search(searchOptions);
     }
 }